Feature : RoomDB : Ensure server data flows through DB data

+++ core/data/datasource/local/src/main/kotlin/com/demo/datausage/core/data/datasource/local/database/entities/DataUsageEntities.kt
@@ -15,5 +15,5 @@ data class MobileDataUsageDB(
     @ColumnInfo(name="quarter")
     val quarter:String,
     @ColumnInfo(name="value")
-    val value:BigDecimal
+    val value:Double
 )
